requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2

# ML & Data Science
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
joblib>=1.4

# Visualization for training/evaluation
matplotlib>=3.9
seaborn>=0.13

# Jupyter Notebook support
notebook>=7.0
ipykernel==6.30.1
